fixed trend refresh issue
